Create brute force
This Python code is used for simulating a brute force attack. Brute force is a type of cyberattack that attempts to guess a password or authentication information by trying all possible combinations.

The main functions of the code are as follows:

brute_force(password, length): Finds the correct password by trying combinations of the specified length.

parallel_brute_force(password, length, processes): Speeds up the brute force process using parallel processes.

guess_password_length(password): Uses brute force to estimate the password's length.

guess_password_complexity(password): Intended to estimate the password's complexity but is not customized.

improved_brute_force(password): Initiates the brute force process by estimating the password's length and complexity.

It's important to emphasize that this code should not be used for illegal or unethical purposes. It should only be used for security testing or educational purposes with appropriate permissions. Unauthorized use of such software is illegal and unethical.